Fix a dependency
Hi,

I hope this can get merge, but understand if it is better to wait for the new registration process.

Neither my REQUIRE file nor my Project.toml indicate a dependency on Test (v0.12.0), yet, for some reason, it is showing up as a dependency here on General.

This is not the end of the world, but slightly annoying because I have 7 other packages (and growing) that depend on this package and all of the manifests are being polluted by this extraneous dependency on Test. I understand they are stdlib dependencies, but still, it would be nice to clean it up if possible.

In any case, I truly appreciate all the hard work and brainpower going into this and I look forward to the new registration process. Thank you 🙏