Pin action to commit hash
Before, everyone with write access to setup-julia could modify the tag and compromise the automerger, thus compromising the entire registry. While the usual settings like enforced 2FA and branch protections are enabled on the action repo, it makes sense to go beyond that for something as key to the Julia ecosystem as the General registry, and minimise the attack surface as much as possible which in this case would mean reducing the amount of GH accounts that could be targetted to gain access to the registry. Additionally, using latest risks that there might be uncaught bugs.

Unlike tags, commit hashes are can't easily be changed, so the potential risk is much lower.