Eliminate `+c` version of OpenSSL_jll.jl
In the saga of continuing to fuzz the `Yggdrasil -> BinaryBuilder -> General` pipeline, we have found that one recipe (OpenSSL) had a build number already specified in the source version number, which took precedence over the build versions we typically use to denote separate runs of the build pipeline.  We've since patched BB to not allow this.