Restore Singular_jll compatibility with Julia 1.6
On the binary level, all versions of Singular_jll v4.1.3 are compatible with both GMP 6.1.2 and 6.2.0, *except* v4.1.3+4 which was GMP >=6.2.0 only, and was yanked. However, due to a bug, v4.1.3+5 specified that it requires precisely GMP_jll 6.1.2. That was wrong, however, and caused by a bug in BinaryBuilderBase.jl which since has been fixed.

This patch now unblocks Singular_jll on Julia 1.6.