Reduce timeout of CI jobs to 60 minutes (#26196)
The default timeout is 360 minutes (6 hours), but I've never seen packages taking more ~10 minutes in the very worst case.  Currently all CI jobs on nightly are hanging, creating a very long queue of jobs, waiting for them to be cancelled.  This change reduces the timeout in order to at least reduce the queue