Change QSymbolics URL
The QSymbolicsBase, QSymbolicsClifford, and QSymbolicsOptics packages
are in the same tree as QuantumSymbolics (they are dependencies of the
latter).

These three packages were registered without issues but then the
registration of QuantumSymbolics originally failed and required a change
of name. That change of name now requires also a change of URL so that
we can continue registering these packages.