Yank date-versioned patchelf build (#85487)
Patchelf originally was built on a version that had nothing to do with any version number, so we applied a date version. However, now Pkg things that version is the latest, so it always chooses that one when you try to install it. To fix this, just drop that old version.