Fix StaticArrays and FillArrays compat bounds for Statistics (#90310)
On Julia 1.1 to 1.3, `Pkg.test` fails for these two packages as
they have been marked to require Statistics 1.10 or below in accordance
to their dependency on Julia 1.0 or below, but only Statistics 1.11
is registered. Registering an older Statistics version is apparently
not enough to fix the problem, probably because of a resolver issue.
Allow Statistics 1.11 as it works.