Adjust Onda's compatibility with TimeSpans
Onda currently defines ArrowTypes interoperability for TimeSpans types, which is a holdover from before TimeSpans became its own package. Said interop will be moved to an ArrowTypes extension for TimeSpans and released as TimeSpans 1.1.0. Onda currently declares compatibility with TimeSpans 1, but in actuality it will be incompatible with TimeSpans 1.1 because the duplicated method definitions will break precompilation. To fix this, we can adjust the declared compatibility bound here in the registry from 1 to 1.0, which means that existing Onda versions won't be compatible with TimeSpans 1.1. A subsequent release of Onda will remove the interop and set its minimum TimeSpans version to 1.1.