Ensure CURL_jll matches LibCURL_jll more precisely
It is not allowable for CURL_jll to get a version of LibCURL_jll that varies too much from its own version.  Note that because LibCURL_jll is a stdlib this is not a problem for most users, but if you're using something like JLLPrefixes.jl to do your package instantiation, you can end up in a situation like this.