Fix repo URLs in non-canonical form (mostly OrdinaryDiffEq)
See the slack discussion. Some dropped the `.git` which makes JuliaRegistrator fail thinking the URL is not found. This seems to come from LocalRegistry.jl uses. This fixes it for all cases in the registry, which is mostly the OrdinaryDiffEq cases, plus one that was found by the regex which is BlockArrays.jl