Transfer packages to WIAS-PDELib organization (#118431)
* Transfer packages to WIAS-PDELib organization

I (@j-fu) and Christian Merdon (@chmerdon) want to transfer a number of packages authored by us to
a new organization WIAS-PDELib. We are both owners of the new organization. In addition,
we want to transfer WIAS-BERLIN/PDELib.jl to that org. (I am one of the owners of WIAS-BERLIN org).

* fix toml in ExtendableFEM