New package: StateSelection
This registers https://github.com/JuliaComputing/StateSelection.jl.
This package implements structural transformation and analysis of
equation systems. This code has bit of history. It started out
as a translation of similar code in Modia.jl in a private package.
It was thereafter open sourced and incorporated in MTK, then split
out again into this package, which is the version used in Cedar's
DAECompiler. The package is currently registered in the Cedar public
registry, but not General. This PR moves the registration to General
to enable other packages in general to depend on it. Hopefully this
will become the canonical version of this code that everybody
just uses.

Two versions are being registered, 0.2.0, and 0.2.1, which are
the two versions that are public in the Cedar registry. There is
also a 0.1.0 version, but this version was never publicly registered,
so it is omitted from this registration as well.