AutoMerge: update staging workflow for 2nd label job, like regular workflow (#140298)
This aligns the staging workflow better to the prod workflow. The prod
workflow has 2 jobs since
https://github.com/JuliaRegistries/General/pull/130112.

This means when someone adds a behavior-affecting label, we run the
separate AutoMerge job to rerun automerge and update the status check
with the new info (eg skipping name similarity). This uses the same
logic for the new `AutoMerge_staging-label:` job. That way we won't get
a red check from the staging workflow when overriding automerge name
similarity, like in
https://github.com/JuliaRegistries/General/pull/140267.

Discussed on slack here:
https://julialang.slack.com/archives/C02174EA421/p1760359829796529?thread_ts=1760359404.837439&cid=C02174EA421